Change cantidad and cantidad_minima data types to decimals

+++ database/migrations/2023_11_22_202453_create_inventarios_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('inventarios', function (Blueprint $table) {
             $table->bigIncrements('id')->nullable(false)->unsigned();
             $table->string('nombre', 50)->nullable(false)->unique();
-            $table->integer('cantidad')->nullable(false);
-            $table->integer('cantidad_minima')->nullable(false);
+            $table->decimal('cantidad', 8, 3)->nullable(false);
+            $table->decimal('cantidad_minima', 8, 3)->nullable(false);
             $table->string('tipo', 100)->nullable(false);
             $table->string('unidad_medida', 10)->nullable(false);
             $table->timestamps();
